requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
attrs==25.4.0
billiard==4.2.2
celery==5.5.3
certifi==2025.10.5
cffi==2.0.0
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==46.0.2
deprecation==2.1.0
Django==5.1.6
django-environ==0.12.0
django-filter==24.3
django-js-asset==3.1.2
django-mptt==0.18.0
djangorestframework==3.16.1
drf-spectacular==0.28.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
msgpack==1.1.1
mysqlclient==2.2.7
packaging==25.0
postgrest==2.21.1
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
pycparser==2.23
pydantic==2.11.10
pydantic_core==2.33.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==4.0.0
PyYAML==6.0.3
realtime==2.21.1
redis==5.2.1
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
storage3==2.21.1
StrEnum==0.4.15
supabase==2.21.1
supabase-auth==2.21.1
supabase-functions==2.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
websockets==15.0.1
wheel==0.45.1
whitenoise==6.11.0
